backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.2.0
pytesseract>=0.3.10
opencv-python>=4.8.0
spacy>=3.7.0
transformers>=4.35.0
torch>=2.1.0
numpy>=1.24.0
pandas>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
geoalchemy2>=0.14.0
httpx>=0.25.0
aiofiles>=23.2.0
google-generativeai>=0.3.0
